implement FilterBase::addBoilerplate()
+ various fixes and other minor missing features of the example extractor

(main part of #948)

+++ src/markdown-extensions/extracted-code-blocks/ExtractFilter.php
@@ -39,6 +39,7 @@ final class ExtractFilter extends FilterBase {
   <<__Override>>
   protected static function processMissingTypecheckerOutput(
     string $hack_file_path,
+    bool $needs_example,
   ): void {
     list($stdout, $_stderr) = \HH\Asio\join(
       ExampleTypechecker\typecheck_example_async($hack_file_path),
@@ -63,15 +64,28 @@ final class ExtractFilter extends FilterBase {
       Files::TYPECHECKER_EXPECT,
       Files::EXAMPLE_TYPECHECKER_OUT,
       Files::TYPECHECKER_EXPECTF,
+      Files::TYPECHECKER_EXPECTREGEX,
+      $needs_example,
     );
   }
 
   <<__Override>>
   protected static function processMissingHHVMOutput(
     string $hack_file_path,
+    bool $needs_example,
   ): void {
+    $config = $hack_file_path.'.'.Files::INI;
+    if (!\file_exists($config)) {
+      $config = LocalConfig::ROOT.'/api-sources/hhvm/hphp/test/config.ini';
+    }
+
     $command = vec[
       \PHP_BINARY,
+      // use the same config as the test runner in ExamplesTest
+      '-c',
+      $config,
+      '-vEval.EnableArgsInBacktraces=false',
+      '-dhhvm.hack.lang.disable_xhp_element_mangling=true',
       '-dauto_prepend_file='.
       LocalConfig::ROOT.
       '/src/utils/_private/init_docs_autoloader.php',
@@ -84,7 +98,7 @@ final class ExtractFilter extends FilterBase {
     $output = null;
     \exec($command.' 2>&1', inout $output, inout $exit_code);
 
-    $match = Regex\first_match($hack_file_path, re"#(.*)/user-documentation/#");
+    $match = Regex\first_match($hack_file_path, re"#(.*/user-documentation)/#");
 
     self::writeOutput(
       Str\join($output, "\n"),
@@ -93,6 +107,8 @@ final class ExtractFilter extends FilterBase {
       Files::HHVM_EXPECT,
       Files::EXAMPLE_HHVM_OUT,
       Files::HHVM_EXPECTF,
+      Files::HHVM_EXPECTREGEX,
+      $needs_example,
     );
   }
 
@@ -109,24 +125,47 @@ final class ExtractFilter extends FilterBase {
     Files $out_file,
     Files $example_file,
     Files $expectf_file,
+    Files $expectregex_file,
+    bool $needs_example,
   ): void {
+    // Ensure newline at the end of the file.
+    if ($output !== '' && !Str\ends_with($output, "\n")) {
+      $output .= "\n";
+    }
+
     $out_path = $hack_file_path.'.'.$out_file;
     $example_path = $hack_file_path.'.'.$example_file;
     $expectf_path = $hack_file_path.'.'.$expectf_file;
+    $expectregex_path = $hack_file_path.'.'.$expectregex_file;
 
-    if (
+    if (\file_exists($expectf_path) || \file_exists($expectregex_path)) {
+      // Only generate example output.
+      invariant(
+        $needs_example && !\file_exists($example_path),
+        'Tried to auto-generate %s but it already exist (or should not exist)!',
+        $example_path,
+      );
+      \file_put_contents(
+        $example_path,
+        $path_to_sanitize is nonnull
+          ? Str\replace($output, $path_to_sanitize, '/home/example')
+          : $output,
+      );
+    } else if (
       $path_to_sanitize is nonnull && Str\contains($output, $path_to_sanitize)
     ) {
       // Note: One of these may have been explicitly specified in the code
       // block, in which case only auto-generate the other one.
       invariant(
-        !\file_exists($example_path) || !\file_exists($expectf_path),
-        'Tried to auto-generate %s and/or %s but both already exist!',
+        ($needs_example && !\file_exists($example_path)) ||
+        !\file_exists($expectf_path),
+        'Tried to auto-generate %s and/or %s but both already exist (or '.
+        'should not exist)!',
         $example_path,
         $expectf_path,
       );
 
-      if (!\file_exists($example_path)) {
+      if ($needs_example && !\file_exists($example_path)) {
         \file_put_contents(
           $example_path,
           Str\replace($output, $path_to_sanitize, '/home/example'),
@@ -137,7 +176,8 @@ final class ExtractFilter extends FilterBase {
         \file_put_contents(
           $expectf_path,
           Str\replace($output, '%', '%%')
-            |> Str\replace($$, $path_to_sanitize, '%s'),
+            |> Str\replace($$, $path_to_sanitize, '%s')
+            |> Regex\replace($$, re"/string\\([0-9]+\\)/", 'string(%d)'),
         );
       }
     } else {
